refactor: fetch store at layout level
